Rate optimize (#185)
* merge with main, pass compile
  * fix tcp rpc adapter

  * update submodule

  * pass compile

  * fixbug

  * pass compile

  * WIP, refactoring frontend stub

  * WIP, refactoring mrpc lib

  * (cherry picked from commit ce2fcc8a4edd568032360bb064a334361e1ffc31)

  * update todo

  * minor update

  * backport null policy

  * remove local_ex in masstree test, improve its performance by a lot

  * format

  * update rpc_bench_rate to make it run a little longer

  * Use PagePadded in ResourceSlab. finally the performance scales linearly

  * save

  * save

  * save

  * multinic and core affinity

  * fix compile and linkage issue

  * update cpu_performance.sh script

  * minor change, use a slab to store the rpc context, latency is increased somehow

  * save, scalability still not ideal

  * WIP, using sharded_slab

  * brutly move concurrent data structures to thread local

  * minor update

* fixbug, forget to wake the future in stub

* format the code

* minor update

* update readme and CI

* fix

+++ src/interface/src/lib.rs
@@ -17,10 +17,10 @@ pub enum Error {
 
 #[repr(transparent)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct Handle(pub u32);
+pub struct Handle(pub u64);
 
 impl Handle {
-    pub const INVALID: Handle = Handle(u32::MAX);
+    pub const INVALID: Handle = Handle(u64::MAX);
 }
 
 pub trait AsHandle {
@@ -52,7 +52,7 @@ pub struct EventChannel(pub Handle);
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub struct CompletionQueue(pub Handle);
 
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
 pub struct ProtectionDomain(pub Handle);
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
